requirements.txt
aiogram>=2.11.2,<3.0
aiohttp==3.7.4
aioredis==1.3.1
python-dotenv
gino
Flask
Jinja2
SQLAlchemy
